chapter_004/src/test/java/ru/kbond/collections/list/SimpleQueueTest.java
package ru.kbond.collections.list;

import org.junit.Test;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

/**
 * Test.
 * @author kbondarenko
 * @since 27.02.2018
 * @version 1
 */
public class SimpleQueueTest {
    /**
     * Test.
     */
    @Test
    public void whenAddIntegersThanGetIntegersFIFO() {
        SimpleQueue<Integer> queue = new SimpleQueue<>();
        queue.push(1);
        queue.push(2);
        queue.push(3);

        assertThat(queue.poll(), is(1));
        assertThat(queue.poll(), is(2));
        assertThat(queue.poll(), is(3));
    }
}